quality-feedback-system/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.23.0
apscheduler>=3.10.0
jinja2>=3.1.0
pyyaml>=6.0
requests>=2.31.0
python-dateutil>=2.8.0
